# server/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "CDC-CBIT API",
    "description": "The following is the list of all the endpoints available for different end-users."
  },
  "host": "localhost:3080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "X-API-KEY",
      "description": "any description..."
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/students/": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "first_name": {
                  "type": "string",
                  "example": "any"
                },
                "last_name": {
                  "type": "string",
                  "example": "any"
                },
                "full_name": {
                  "type": "string",
                  "example": "any"
                },
                "roll_number": {
                  "type": "string",
                  "example": "any"
                },
                "branch": {
                  "type": "string",
                  "example": "any"
                },
                "section": {
                  "type": "string",
                  "example": "any"
                },
                "placement_batch": {
                  "type": "string",
                  "example": "any"
                },
                "primary_email": {
                  "type": "string",
                  "example": "any"
                },
                "secondary_email": {
                  "type": "string",
                  "example": "any"
                },
                "mobile": {
                  "type": "string",
                  "example": "any"
                },
                "secondary_mobile": {
                  "type": "string",
                  "example": "any"
                },
                "cgpa": {
                  "type": "string",
                  "example": "any"
                },
                "backlogs": {
                  "type": "string",
                  "example": "any"
                },
                "inter_or_diploma_college": {
                  "type": "string",
                  "example": "any"
                },
                "inter_or_diploma_percentage": {
                  "type": "string",
                  "example": "any"
                },
                "school": {
                  "type": "string",
                  "example": "any"
                },
                "school_percentage": {
                  "type": "string",
                  "example": "any"
                },
                "eamcet_rank": {
                  "type": "string",
                  "example": "any"
                },
                "jee_rank": {
                  "type": "string",
                  "example": "any"
                },
                "parent_name": {
                  "type": "string",
                  "example": "any"
                },
                "address": {
                  "type": "string",
                  "example": "any"
                },
                "city": {
                  "type": "string",
                  "example": "any"
                },
                "state": {
                  "type": "string",
                  "example": "any"
                },
                "zipcode": {
                  "type": "string",
                  "example": "any"
                },
                "gender": {
                  "type": "string",
                  "example": "any"
                },
                "date_of_birth": {
                  "type": "string",
                  "example": "any"
                },
                "photo_url": {
                  "type": "string",
                  "example": "any"
                },
                "resume_url": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "is_verified": {
                  "type": "string",
                  "example": "any"
                },
                "first_name": {
                  "type": "string",
                  "example": "any"
                },
                "last_name": {
                  "type": "string",
                  "example": "any"
                },
                "full_name": {
                  "type": "string",
                  "example": "any"
                },
                "roll_number": {
                  "type": "string",
                  "example": "any"
                },
                "branch": {
                  "type": "string",
                  "example": "any"
                },
                "section": {
                  "type": "string",
                  "example": "any"
                },
                "placement_batch": {
                  "type": "string",
                  "example": "any"
                },
                "primary_email": {
                  "type": "string",
                  "example": "any"
                },
                "secondary_email": {
                  "type": "string",
                  "example": "any"
                },
                "mobile": {
                  "type": "string",
                  "example": "any"
                },
                "secondary_mobile": {
                  "type": "string",
                  "example": "any"
                },
                "cgpa": {
                  "type": "string",
                  "example": "any"
                },
                "backlogs": {
                  "type": "string",
                  "example": "any"
                },
                "inter_or_diploma_college": {
                  "type": "string",
                  "example": "any"
                },
                "inter_or_diploma_percentage": {
                  "type": "string",
                  "example": "any"
                },
                "school": {
                  "type": "string",
                  "example": "any"
                },
                "school_percentage": {
                  "type": "string",
                  "example": "any"
                },
                "eamcet_rank": {
                  "type": "string",
                  "example": "any"
                },
                "jee_rank": {
                  "type": "string",
                  "example": "any"
                },
                "parent_name": {
                  "type": "string",
                  "example": "any"
                },
                "address": {
                  "type": "string",
                  "example": "any"
                },
                "city": {
                  "type": "string",
                  "example": "any"
                },
                "state": {
                  "type": "string",
                  "example": "any"
                },
                "zipcode": {
                  "type": "string",
                  "example": "any"
                },
                "gender": {
                  "type": "string",
                  "example": "any"
                },
                "date_of_birth": {
                  "type": "string",
                  "example": "any"
                },
                "photo_url": {
                  "type": "string",
                  "example": "any"
                },
                "resume_url": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/students/myProfile": {
      "get": {
        "tags": [],
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/students/{userId}": {
      "delete": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/students/verifyStudents/{stId}": {
      "patch": {
        "tags": [],
        "description": "",
        "parameters": [
          {
            "name": "stId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
...
}